Improve error handling
Avoid returning the array of errors from the public methods and instead
return null or a single instance of Error that contains one or more
instances of Error in it. Also, reject a promise returned from
`metaschema.fs.load()` on errors during `Metaschema` object creation
instead of resolving it with an array of errors.

PR-URL: https://github.com/metarhia/metaschema/pull/215